seperated sign up form in 3 components
