+++ src/app/_components/SignUp/Interests.tsx
@@ -1,7 +1,108 @@
 import React from "react";
+import { useContentFormInterest } from "~/app/_utils/Hooks/useContentForm";
+import { Controller } from "react-hook-form";
 
-const Interests = () => {
-  return <div>Interests</div>;
+interface InterestsFormProps {
+  onNext: () => void; // Expecting a function as a prop
+}
+
+const interestsIndex = [
+  "Animals",
+  "Art",
+  "Books",
+  "Comedy",
+  "Culture",
+  "Education",
+  "Food",
+  "Health",
+  "History",
+  "Music",
+  "Nature",
+  "News",
+  "Politics",
+  "Science",
+  "Sports",
+  "Technology",
+  "Travel",
+  "TV",
+  "Video Games",
+];
+
+const Interests: React.FC<InterestsFormProps> = ({ onNext }) => {
+  const {
+    register: registerInterest,
+    handleSubmit: handleSubmitInterest,
+    trigger: triggerInterest,
+    getValues: getValuesInterest,
+    watch: watchInterest,
+    setValue: setValueInterest,
+    control: controlInterest,
+  } = useContentFormInterest();
+
+  const handleNext = async (data) => {
+    const isValid = await triggerInterest();
+    if (isValid) {
+      onNext(data);
+    }
+  };
+
+  return (
+    <form onSubmit={handleSubmitInterest(onNext)} className="w-full">
+      <div className="w-full px-10">
+        <div className="mb-3 flex flex-col">
+          <h1 className="text-3xl font-medium text-font-primary">
+            What are your interests?
+          </h1>
+          <h1 className="text-lg font-medium text-font-secondary">
+            We will use this to help customize the experience
+          </h1>
+          <div className="relative mb-[200px] mt-6 flex h-auto w-full flex-wrap gap-2">
+            <Controller
+              name="interests"
+              control={controlInterest}
+              defaultValue={[]}
+              render={({ field: { onChange, value }, fieldState }) => (
+                <>
+                  {interestsIndex.map((interest) => (
+                    <button
+                      key={interest}
+                      type="button"
+                      onClick={() => {
+                        const newValue = value.includes(interest)
+                          ? value.filter((item) => item !== interest)
+                          : [...value, interest];
+                        onChange(newValue);
+                      }}
+                      className={`pill inline-block rounded-full bg-gray-200 px-4 py-2 text-sm text-gray-800 ${
+                        value.includes(interest)
+                          ? "border border-brand-primary bg-[#fafbff] text-brand-primary"
+                          : "bg-gray-100 text-font-primary"
+                      }`}
+                    >
+                      {interest} {value.includes(interest) && "✕"}
+                    </button>
+                  ))}
+                  {fieldState.error && (
+                    <span className="absolute -bottom-20 -right-10 text-red-500">
+                      {fieldState.error.message}
+                    </span>
+                  )}
+                </>
+              )}
+            />
+          </div>
+          <div className="mt-4 flex items-center justify-center">
+            <button
+              type="submit" // Use type="submit" for form submission
+              className="w-1/3 rounded-sm bg-brand-primary px-3 py-2 text-lg text-white"
+            >
+              Next
+            </button>
+          </div>
+        </div>
+      </div>
+    </form>
+  );
 };
 
 export default Interests;
